Refactor cleanup, docstrings, etc (+16) Squashed commits: Fix type for argument [1a69bd1] Homogenize language /s/pipeline/workflow [9325f68] Remove unused functions [f0e537a] Add missing docstrings and cleanup comments [f1787cb] Update DAG image [6d5eb82] Remove old test and fix config options [2f13248] Add hera-index s3 link to config [da568a0] Remove old toil-lib depedency [9acd812] Cast cores requirement as string [3507eb0] Remove BamQC from config [5df7c01] Fix file name for sorted output bam [bd64bbb] Add missing args to run_kallisto and run_hera [469fc43] Encapsulate `wrapJobFn`s which create multiple children internally [56b6357] Add .gz attribute to config for input fastqs [796be07] Replace Expando call with recursive expando [29a388a] Add recursive expando for handling dashes in input names [d6de3ba] Resolve circular imports (+1 squashed commit) Squashed commits: [5b542f8] Update dependency for Toil
Using new features like job.log() and job.tempDir

+++ src/toil_rnaseq/utils/files.py
@@ -1,9 +1,6 @@
 import os
 import shutil
 import tarfile
-from contextlib import closing
-
-from toil_lib import require
 
 
 def tarball_files(tar_name, file_paths, output_dir='.', prefix=''):
@@ -37,37 +34,6 @@ def __forall_files(file_paths, output_dir, op):
         op(file_path, dest)
 
 
-def copy_file_job(job, name, file_id, output_dir):
-    """
-    Job version of move_files for one file
-
-    :param JobFunctionWrappingJob job: passed automatically by Toil
-    :param str name: Name of output file (including extension)
-    :param str file_id: FileStoreID of file
-    :param str output_dir: Location to place output file
-    """
-    work_dir = job.fileStore.getLocalTempDir()
-    fpath = job.fileStore.readGlobalFile(file_id, os.path.join(work_dir, name))
-    copy_files([fpath], output_dir)
-
-
-def move_files(file_paths, output_dir):
-    """
-    Moves files from the working directory to the output directory.
-
-    Important note: this function can couple dangerously with caching.
-    Specifically, if this function is called on a file in the cache, the cache
-    will contain a broken reference. This may lead to a non-existent file path
-    being passed to later jobs. Don't call this function on files that are in
-    the cache, unless you know for sure that the input file will not be used by
-    any later jobs.
-
-    :param str output_dir: Output directory
-    :param list[str] file_paths: Absolute file paths to move
-    """
-    __forall_files(file_paths, output_dir, shutil.move)
-
-
 def copy_files(file_paths, output_dir):
     """
     Moves files from the working directory to the output directory.
@@ -78,44 +44,34 @@ def copy_files(file_paths, output_dir):
     __forall_files(file_paths, output_dir, shutil.copy)
 
 
-def consolidate_tarballs_job(job, fname_to_id):
+def move_files(file_paths, output_dir):
     """
-    Combine the contents of separate tarballs into one.
-    Subdirs within the tarball will be named the keys in **fname_to_id
-
-    :param JobFunctionWrappingJob job: passed automatically by Toil
-    :param dict[str,str] fname_to_id: Dictionary of the form: file-name-prefix=FileStoreID
-    :return: The file store ID of the generated tarball
-    :rtype: str
+    Moves files from the working directory to the output directory.
+    Important note: this function can couple dangerously with caching.
+    Specifically, if this function is called on a file in the cache, the cache
+    will contain a broken reference. This may lead to a non-existent file path
+    being passed to later jobs. Don't call this function on files that are in
+    the cache, unless you know for sure that the input file will not be used by
+    any later jobs.
+    :param str output_dir: Output directory
+    :param list[str] file_paths: Absolute file paths to move
     """
-    work_dir = job.fileStore.getLocalTempDir()
-    # Retrieve output file paths to consolidate
-    tar_paths = []
-    for fname, file_store_id in fname_to_id.iteritems():
-        p = job.fileStore.readGlobalFile(file_store_id, os.path.join(work_dir, fname + '.tar.gz'))
-        tar_paths.append((p, fname))
-    # I/O
-    # output_name is arbitrary as this job function returns a FileStoreId
-    output_name = 'foo.tar.gz'
-    out_tar = os.path.join(work_dir, output_name)
-    # Consolidate separate tarballs into one
-    with tarfile.open(os.path.join(work_dir, out_tar), 'w:gz') as f_out:
-        for tar, fname in tar_paths:
-            with tarfile.open(tar, 'r') as f_in:
-                for tarinfo in f_in:
-                    with closing(f_in.extractfile(tarinfo)) as f_in_file:
-                        tarinfo.name = os.path.join(output_name, fname, os.path.basename(tarinfo.name))
-                        f_out.addfile(tarinfo, fileobj=f_in_file)
-    return job.fileStore.writeGlobalFile(out_tar)
+    __forall_files(file_paths, output_dir, shutil.move)
 
 
 def generate_file(file_path, generate_func):
     """
     Checks file existance, generates file, and provides message
+
     :param str file_path: File location to generate file
     :param func generate_func: Function used to generate file
+    :return: Path to file
+    :rtype: str
     """
-    require(not os.path.exists(file_path), file_path + ' already exists!')
-    with open(file_path, 'w') as f:
-        f.write(generate_func())
-    print('\t{} has been generated in the current working directory.'.format(os.path.basename(file_path)))
+    if os.path.exists(file_path):
+        print('File "{}" already exists! Doing nothing.'.format(file_path))
+    else:
+        with open(file_path, 'w') as f:
+            f.write(generate_func())
+        print('\t{} has been generated in the current working directory.'.format(os.path.basename(file_path)))
+    return file_path
